@@ -1,94 +1,38 @@
-import React from 'react';
-import { NerdGraphQuery } from 'nr1';
+import React, { useContext, useEffect } from 'react';
+import { PlatformStateContext, nerdlet, AutoSizer } from 'nr1';
+import { ProvideData } from '../../src/context/data';
+import AccountMaturity from './AccountMaturity';
+import { Messages } from '@newrelic/nr-labs-components';
 
-import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
-import {
-  OverviewPanel,
-  APMPanel,
-  BrowserPanel,
-  SynthPanel,
-  InfraPanel,
-  KubernetesPanel,
-  InsightsPanel,
-  LogPanel,
-  ProgramPanel,
-  MobilePanel,
-  WorkloadPanel,
-  SLMPanel,
-  NPMPanel
-} from 'maturity-products/dist/entities';
-import {
-  ApplicationCtxProvider,
-  MaturityScoreCtxProvider
-} from 'maturity-products/dist/contexts';
+export default function AccountMaturityRoot() {
+  const platformContext = useContext(PlatformStateContext);
 
-export default class MaturityApplication extends React.Component {
-  constructor(props) {
-    super(props);
-  }
+  useEffect(() => {
+    nerdlet.setConfig({
+      accountPicker: true,
+      timePicker: false,
+      actionControls: true
+    });
+  }, []);
 
-  render() {
-    return (
-      <Tabs forceRenderTabPanel>
-        <TabList>
-          <Tab>Overview</Tab>
-          <Tab>APM</Tab>
-          <Tab>Browser</Tab>
-          <Tab>Synthetics</Tab>
-          <Tab>Infrastructure</Tab>
-          <Tab>Kubernetes</Tab>
-          <Tab>Insights</Tab>
-          <Tab>Log</Tab>
-          <Tab>Programmability</Tab>
-          <Tab>Mobile</Tab>
-          <Tab>Workloads</Tab>
-          <Tab>SLM</Tab>
-          <Tab>NPM</Tab>
-        </TabList>
-        <ApplicationCtxProvider nr1graph={NerdGraphQuery}>
-          <MaturityScoreCtxProvider>
-            <TabPanel>
-              <OverviewPanel />
-            </TabPanel>
-            <TabPanel>
-              <APMPanel />
-            </TabPanel>
-            <TabPanel>
-              <BrowserPanel />
-            </TabPanel>
-            <TabPanel>
-              <SynthPanel />
-            </TabPanel>
-            <TabPanel>
-              <InfraPanel />
-            </TabPanel>
-            <TabPanel>
-              <KubernetesPanel />
-            </TabPanel>
-            <TabPanel>
-              <InsightsPanel />
-            </TabPanel>
-            <TabPanel>
-              <LogPanel />
-            </TabPanel>
-            <TabPanel>
-              <ProgramPanel />
-            </TabPanel>
-            <TabPanel>
-              <MobilePanel />
-            </TabPanel>
-            <TabPanel>
-              <WorkloadPanel />
-            </TabPanel>
-            <TabPanel>
-              <SLMPanel />
-            </TabPanel>
-            <TabPanel>
-              <NPMPanel />
-            </TabPanel>
-          </MaturityScoreCtxProvider>
-        </ApplicationCtxProvider>
-      </Tabs>
-    );
-  }
+  return (
+    <div>
+      <Messages
+        org="newrelic"
+        repo="nr1-account-maturity"
+        branch="main"
+        directory="src/"
+        fileName="messages"
+        timeoutPeriod={8 * 7 * 24 * 60 * 60} // 8 weeks
+      />
+
+      <ProvideData platformContext={platformContext}>
+        <AutoSizer>
+          {({ width, height }) => (
+            <AccountMaturity width={width} height={height} />
+          )}
+        </AutoSizer>
+      </ProvideData>
+    </div>
+  );
 }
