@@ -1,4 +1,4 @@
 module.exports = {
-    singleQuote: true,
-    trailingComma: "es5",
+  singleQuote: true,
+  trailingComma: 'none'
 };
